% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{fill_missing}
\alias{fill_missing}
\title{Replace missing values in a column}
\usage{
fill_missing(.data, ...)
}
\arguments{
\item{.data}{A table, or a pipeline.}

\item{...}{`column = value` pairs.}
}
\value{
A pipeline.
}
\description{
Replace missing values in a column
}
