% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{god_sql}
\alias{god_sql}
\title{The query a pipeline becomes}
\usage{
god_sql(pipeline, columns)
}
\arguments{
\item{pipeline}{The pipeline, as text.}

\item{columns}{The table's columns, as `name:type` separated by commas.}
}
\value{
The query, as text.
}
\description{
The query a pipeline becomes
}
